backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
PyJWT==2.8.0
bcrypt==4.0.1
Flask-SocketIO==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
opencv-python==4.8.0.76
numpy>=1.26.0
Pillow==10.0.0 